+++ src/errors.rs
@@ -1,35 +1,11 @@
-use std::error::Error;
-use std::fmt;
+use thiserror::Error;
 
-#[derive(Debug)]
-pub struct InvalidDurationError {}
-
-impl fmt::Display for InvalidDurationError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Invalid duration, possible suffixes: d, w")
-    }
-}
-
-impl Error for InvalidDurationError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        None
-    }
-}
-
-#[derive(Debug)]
-pub struct InvalidChannelConfigError {}
-
-impl fmt::Display for InvalidChannelConfigError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Invalid channel configuration, format: $name:$duration,..."
-        )
-    }
-}
-
-impl Error for InvalidChannelConfigError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        None
-    }
+#[derive(Error, Debug)]
+pub enum ParseChannelConfigError {
+    #[error("`{0}` is not a valid duration suffix, valid suffixes are: d, w")]
+    InvalidDurationSuffix(char),
+    #[error("a duration needs a suffix, valid suffixes are: d, w")]
+    NoDurationSuffix,
+    #[error("invalid format, should be $CHANNEL_NAME:$DURATION,...")]
+    InvalidFormat,
 }