+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Clippy check job
 - Security audit job (runs overnight)
-- `failure` crate for simpler Results and context
+- `failure` crate for simpler results and context
+- `thiserror` crate for easier error creation
 ### Changed
 - Unified GitHub Actions workflow