+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2020-12-08
 ### Added
 - Clippy check job
 - Security audit job (runs overnight)
@@ -24,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic GitHub actions setup
 
 [Unreleased]: https://github.com/bahlo/discord-retention-bot/compare/v0.1.0...HEAD
+[1.0.0]: https://github.com/bahlo/discord-retention-bot/compare/v0.1.0...v1.0.0
 [0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.1.0